Add support for HTTP
When enabled, serve a self-updating status widget via HTTP which can be
included in OBS using a browser source.

The widget writes its default HTML and JavaScript to the application's
working directory (like the log file) where they can be customized by the
user. New files will be written (and edited files re-read) only when
connecting to the sensor, so in order to use updated files, it's necessary
to disconnect and re-connect the sensor.